the full redirection process seems to work considering if 2fa or not
