+++ frontend/package.json
@@ -15,6 +15,7 @@
     "react-auth-code-input": "^3.2.1",
     "react-dom": "^18.2.0",
     "react-qr-code": "^2.0.11",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
