+++ frontend/src/App.tsx
@@ -5,343 +5,27 @@ import FormControlLabel from '@mui/material/FormControlLabel';
 import React from 'react';
 import { useState, useEffect } from 'react';
 import axios from 'axios';
-// import QRCode from 'react-qr-code';
 import AuthCode from 'react-auth-code-input';
-// import { useRef } from 'react';
+import verify2FA from './components/2fa-verify';
+import LoginPage from './components/login';
+import HomePage from './components/homepage';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
+import { Routes } from 'react-router-dom';
 
 function App() {
   
-  const [checked, setChecked] = React.useState(false);
-  const [twoFAActivated, setTwoFAActivated] = React.useState(false);
-  const [qrcodeDataUrl, setQrcodeDataUrl] = React.useState('');
-  const [token, setToken] = useState('');
-  const [twoFACode, setTwoFACode] = React.useState('');
-
-  useEffect(() => {
-    if (token !== ''){
-      console.log("token: ", token);
-      check2FAStatus(token).then((status:any) => status.json()).then((status:any) => {
-        console.log("status: ", status);
-        setChecked(status.twoFactorAuth);
-      });
-    }
-    let cookieToken = document.cookie.replace(/(?:(?:^|.*;\s*)token\s*\=\s*([^;]*).*$)|^.*$/, "$1");
-    if (cookieToken) {
-      setToken(cookieToken);
-    }
-
-  }, [token]);
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-
-    // if (!checked === true){
-      setChecked(event.target.checked);
-      console.log("status: ", !checked);
-    // }
-
-      axios.post('http://localhost:3333/auth/2fa/enable', { token, twoFactorAuth: !checked }).then(response => {
-      // setQrcodeDataUrl(response.data);
-      // console.log("qrcodeDataUrl: ", response.data);
-
-    // }
-
-    console.log(response);
-  }).catch(error => {
-      console.error(error);
-    });
-  };
-
-  /**
-  ** Display a base64 URL inside an iframe in another window.
-  **/
-  function debugBase64(base64URL: string){
-    var win = window.open();
-    if (win)
-      win.document.write('<iframe src="' + base64URL  + '" frameborder="0" style="border:0; top:0px; left:0px; bottom:0px; right:0px; width:100%; height:100%;" allowfullscreen></iframe>');
-  }
-
-  // function handleQRCodeLoad() {
-  //   console.log("handle qr load url -> ", qrcodeDataUrl);
-  //   debugBase64(qrcodeDataUrl);
-  // };
-
-  const handleOnChange = (code: string) => {
-    setTwoFACode(code);
-    console.log("2fa code: ", code);
-  };
-
-  async function generateQRCode(): Promise<any> {
-    try {
-      const response = await fetch('http://localhost:3333/auth/2fa/generate', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({ token: token })
-      });
-      const data = await response.json();
-      console.log(data);
-      setQrcodeDataUrl(data);
-      debugBase64(data);
-      return data;
-    } catch (error) {
-      console.error(error);
-      // handle error
-    }
-  }
-
-  async function activate2FA(): Promise<any> {
-    const response = await fetch('http://localhost:3333/auth/2fa/verify', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ twoFACode: twoFACode })
-    });
-    const data = await response.json();
-    if (data)
-    {
-      // setChecked(true);
-      setTwoFAActivated(true);
-      // axios.post('http://localhost:3333/auth/2fa/enable', { token, twoFactorAuth: twoFAActivated }).then(response => {
-      // setQrcodeDataUrl(response.data);
-      // console.log("qrcodeDataUrl: ", response.data);
-  
-      // }
-  
-      // console.log(response);
-    // }).catch(error => {
-        // console.error(error);
-      // });
-      window.location.href = "http://localhost:3000/mainpage";
-    }
-    console.log(data);
-    return data;
-  }
-  
-  //submit2FACode function is not used right now, but keeping it..
-  async function submit2FACode(): Promise<any> {
-    const response = await fetch('http://localhost:3333/auth/2fa/verify', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({ twoFACode: twoFACode })
-    });
-    const data = await response.json();
-    if (data)
-    {
-      // setChecked(true);
-      setTwoFAActivated(true);
-      axios.post('http://localhost:3333/auth/2fa/enable', { token, twoFactorAuth: twoFAActivated }).then(response => {
-      setQrcodeDataUrl(response.data);
-      console.log("qrcodeDataUrl: ", response.data);
-  
-      // }
-  
-      // console.log(response);
-    }).catch(error => {
-        console.error(error);
-      });
-      window.location.href = "http://localhost:3000/mainpage";
-    }
-    console.log(data);
-    return data;
-  }
-
-  async function check2FAStatus(accessToken: string): Promise<any>{
-    try {
-      const response = await fetch('http://localhost:3333/auth/2fa/status', {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-          'Authorization': `${accessToken}`
-        },
-      });
-      // const response = await axios.get('http://localhost:3333/auth/2fa/status'
-      return response;
-    } catch (error) {
-
-      console.error(error);
-      // handle error
-    }
-  }
-  
-  if (token) {
     return (
-      <div className="App">
-      <>
-         <Button 
-          text="Logout"
-          onClick={() => {
-            window.open('http://localhost:3333/auth/42/logout', "_self");
-          }}
-        />
-
-         <FormControlLabel control={
-          <Switch
-            checked={checked}
-            onChange={handleChange}
-            inputProps={{ "aria-label": "controlled" }}
-          />
-        } label="Enable 2FA" />
-
-        <p>Token: {token}</p>
-
-        {checked && (
-          <div>
-          {qrcodeDataUrl && (
-            <div>
-                <p>{qrcodeDataUrl}</p>
-            </div>
-          )}
-          </div>
-          )
-        }
-
-        <>
-          <button onClick={generateQRCode}>Generate QR code</button>
-        </>
-
-      <>
-        <AuthCode
-          allowedCharacters='numeric'
-          onChange={handleOnChange}
-        />
-        <button onClick={activate2FA}>Submit code</button>
-      </>
-
-      </>
-
-{/* 
-      {checked && (
-          <div>
-          {qrcodeDataUrl && (
-            <div>
-                <p>{qrcodeDataUrl}</p>
-            </div>
-          )}
-          </div>
-          )
-        } */}
-    </div>
+      <Router>
+        <div>
+          <Routes>
+            <Route path="/" element={<LoginPage/>}/>
+            <Route path="/login" element={<LoginPage/>}/>
+            {/* <Route path="/2fa/verification" element={<verify2FA/>}/> */}
+            <Route path="/homepage" element={<HomePage/>}/>
+          </Routes>
+        </div>
+      </Router>
     );
-  }
-  else if(token && twoFAActivated) {
-    return (
-      <>
-        {qrcodeDataUrl && (
-          <div>
-            <button onClick={() => debugBase64(qrcodeDataUrl)}>Get QR</button>
-             {/* <Button
-                  text="Get QR"
-                  onClick={debugBase64(qrcodeDataUrl)}/> */}
-          </div>
-          )
-        }
-        <AuthCode
-          allowedCharacters='numeric'
-          onChange={handleOnChange}
-        />
-        <button onClick={submit2FACode}>Submit code</button>
-      </>
-    )
-  }
-  else {
-    return (
-            <>
-              <Button 
-              text="Log in with 42"
-              onClick={() => {
-                window.open('https://api.intra.42.fr/oauth/authorize?client_id=u-s4t2ud-c3680374c7c94850b80d768576ab99300705487e1f5c7f758876aaf8fbf5fbdb&redirect_uri=http%3A%2F%2Flocalhost%3A3333%2Fauth%2F42%2Fcallback&response_type=code', "_self");
-              }}
-              />
-            </>
-    );
-  }
-  // return (
-  //   <div className="App">
-
-  //   {!token && (
-  //     <>
-  //       <Button 
-  //       text="Log in with 42"
-  //       onClick={() => {
-  //         window.open('https://api.intra.42.fr/oauth/authorize?client_id=u-s4t2ud-c3680374c7c94850b80d768576ab99300705487e1f5c7f758876aaf8fbf5fbdb&redirect_uri=http%3A%2F%2Flocalhost%3A3333%2Fauth%2F42%2Fcallback&response_type=code', "_self");
-  //       }}
-  //       />
-  //     </>
-  //     )
-  //   }
-
-  //   {token && (
-  //     <>
-  //       <Button 
-  //         text="Logout"
-  //         onClick={() => {
-  //           window.open('http://localhost:3333/auth/42/logout', "_self");
-  //         }}
-  //       />
-
-  //       <FormControlLabel control={
-  //         <Switch
-  //           checked={checked}
-  //           onChange={handleChange}
-  //           inputProps={{ "aria-label": "controlled" }}
-  //         />
-  //       } label="Enable 2FA" />
-
-  //       <p>Token: {token}</p>
-
-  //       {checked && (
-  //         <div>
-  //         {qrcodeDataUrl && (
-  //           <div>
-  //               <p>{qrcodeDataUrl}</p>
-  //           </div>
-  //         )}
-  //         </div>
-  //         )
-  //       }
-
-  //       {qrcodeDataUrl && (
-  //         <div>
-  //            <Button
-  //                 text="Get QR"
-  //                 onClick={debugBase64(qrcodeDataUrl)}/>
-  //         </div>
-  //         )
-  //       }
-
-  //     </>
-  //     )
-  //   }
-
-  //   {token && (
-  //     <>
-  //     {!checked && (
-  //       <>
-  //         <AuthCode
-  //           allowedCharacters='numeric'
-  //           onChange={handleOnChange}
-  //         />
-  //         <button onClick={submit2FACode}>Submit code</button>
-  //       </>
-  //       )
-  //     }
-  //     </>
-  //   )
-  //   }
-  //     <>
-  //       {/* <AuthCode
-  //         allowedCharacters='numeric'
-  //         onChange={handleOnChange}
-  //       />
-  //       <button onClick={submit2FACode}>Submit code</button> */}
-  //     </>
-
-  //   </div>
-  // );
 }
 
 export default App;
